# api/openapi/memory.v1.yaml
openapi: 3.0.3
info:
  title: Agent Memory API
  version: 1.0.0
paths:
...
  /v1/memories/{memory_id}:
    get:
      summary: Get memory
      parameters:
        - name: memory_id
          in: path
          required: true
          schema: { type: string }
        - name: tenant_id
          in: query
          schema: { type: string }
      responses:
        '200': { description: Memory }
    delete:
      summary: Delete memory
      parameters:
        - name: memory_id
          in: path
          required: true
          schema: { type: string }
        - name: tenant_id
          in: query
          schema: { type: string }
      responses:
        '200': { description: Deletion receipt }
...
  /v1/agentic/specs:
    post:
      summary: Create a compact SPEC artifact stored as durable memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecRequest'
      responses:
        '201': { description: Created spec }
...
  /v1/agentic/plans:
    post:
      summary: Create a PLAN artifact stored as durable memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanRequest'
      responses:
        '201': { description: Created plan }
...
components:
  schemas:
...
    ScopeFields:
      type: object
      properties:
        tenant_id: { type: string }
        user_id: { type: string }
        agent_id: { type: string }
        project_id: { type: string }
        session_id: { type: string }
...
    SpecRequest:
      allOf:
        - $ref: '#/components/schemas/ScopeFields'
        - type: object
          required: [title, goal]
          properties:
            title: { type: string }
            goal: { type: string }
            background: { type: string }
            constraints: { type: array, items: { type: string } }
            acceptance_criteria: { type: array, items: { type: string } }
            non_goals: { type: array, items: { type: string } }
            risks: { type: array, items: { type: string } }
            evidence: { type: array, items: { type: string } }
    PlanStep:
      type: object
      required: [title, description]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        files: { type: array, items: { type: string } }
        verification: { type: array, items: { type: string } }
        expected_evidence: { type: array, items: { type: string } }
        status: { type: string }
    PlanRequest:
      allOf:
        - $ref: '#/components/schemas/ScopeFields'
        - type: object
          required: [title, objective, steps]
          properties:
            spec_id: { type: string }
            title: { type: string }
            objective: { type: string }
            steps: { type: array, items: { $ref: '#/components/schemas/PlanStep' } }
            checkpoints: { type: array, items: { type: string } }
            token_budget: { type: integer }
            risk: { type: string, enum: [low, medium, high, critical] }
            evidence_refs: { type: array, items: { type: string } }
...